Find the related block by id
